backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
alembic
python-dotenv
fastapi-limiter
redis
httpx[http2]
tenacity
selectolax
playwright
phonenumbers
readability-lxml
lxml
openai
google-generativeai
